fix: show import match score and split the import flow module
- popup: render import.result matchScore as a "— NN% fit." suffix on the
  success/status-unchanged lines (was dead-ends client-side)
- align extension-protocol.ts's stale matchScore docblock with the
  constants-file doc (best-effort keyword-only, omitted on failure)
- match_live: split score_import_posting_bounded's logging so a genuine
  timeout warns while an ordinary no-resume/no-text/scoring-failure None
  only debugs; add a #fragment cache-key parity variant
- popup: add a doCheckFit test covering the per-connection throttle reply
- extension_bridge: move the import.request flow (ImportOk/result_reply/
  persist_import_application/usable/handle_import) into a new
  import_flow.rs sibling module so mod.rs clears the R8 LOC cap
  (1399 -> 1121), behavior-identical relocation
